refactor: update worker import URLs in markdown plugins
